+++ python/sglang/srt/managers/scheduler_components/pool_stats_observer.py
@@ -11,7 +11,7 @@ from typing import (
     Tuple,
 )
 
-from sglang.srt.mem_cache.multi_ended_allocator import (
+from sglang.srt.mem_cache.allocator.unified_hybrid_swa import (
     UnifiedMambaSWATokenToKVPoolAllocator,
 )
