+++ test/registered/unit/mem_cache/test_unified_byte_accounting.py
@@ -42,7 +42,7 @@ from test_multi_ended_allocator import (
     TestUnifiedSWATokenToKVPoolAllocator as _SwaFixture,
 )
 
-from sglang.srt.mem_cache import multi_ended_allocator as mea
+from sglang.srt.mem_cache.allocator import unified_sub_pool as mea
 from sglang.test.ci.ci_register import register_cpu_ci
 
 register_cpu_ci(est_time=15, suite="base-a-test-cpu")
